requirements.txt
adal==1.2.7
aiofiles==22.1.0
aiohttp==3.8.3
aiokafka==0.8.0
aiosignal==1.3.1
alembic==1.8.1
altair==4.2.0
anyio==3.6.2
argcomplete==2.0.0
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.26.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.2.1
azure-mgmt-storage==20.1.0
azureml-core==1.47.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
blinker==1.5
boto3==1.26.14
botocore==1.29.14
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.5
cryptography==38.0.3
cycler==0.11.0
databricks-cli==0.17.3
decorator==5.1.1
distlib==0.3.6
docker==6.0.1
entrypoints==0.4
fastapi==0.87.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.37.4
frozenlist==1.3.3
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.9
GitPython==3.1.28
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpcio==1.51.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==4.16.0
kafka-python==2.0.2
keyboard==0.13.5
kiwisolver==1.4.4
knack==0.10.0
kubernetes==25.3.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mlflow==2.0.1
mlserver==1.2.0rc4
mlserver-mlflow==1.2.0rc4
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.2
ndg-httpsclient==0.5.1
numba==0.56.4
numpy==1.23.3
oauthlib==3.2.2
orjson==3.8.2
packaging==21.3
pandas==1.5.0
paramiko==2.12.0
pathspec==0.10.2
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.4
portalocker==2.6.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
protobuf==3.20.3
py-grpc-prometheus==0.7.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b3
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.18.1
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-rapidjson==1.9
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.1
requests-auth-aws-sigv4==0.7
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
SecretStorage==3.3.3
semver==2.13.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
starlette==0.21.0
starlette-exporter==0.14.0
streamlit==1.13.0
tabulate==0.9.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
tritonclient==2.27.0
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.20.0
validators==0.20.0
virtualenv==20.16.7
waitress==2.1.2
watchdog==2.1.9
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
zope.event==4.5.0
zope.interface==5.5.2
